# docs/openapi/openapi.v1.json
{"components":{"schemas":{"ArtifactListResponse":{"properties":{"artifacts":{"items":{"$ref":"#/components/schemas/ArtifactOut"},"title":"Artifacts","type":"array"}},"title":"ArtifactListResponse","type":"object"},"ArtifactOut":{"properties":{"expires_at":{"title":"Expires At","type":"string"},"format":{"title":"Format","type":"string"},"height":{"title":"Height","type":"integer"},"id":{"title":"Id","type":"string"},"item_index":{"title":"Item Index","type":"integer"},"s3_key":{"title":"S3 Key","type":"string"},"seed":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Seed"},"url":{"title":"Url","type":"string"},"width":{"title":"Width","type":"integer"}},"required":["id","format","width","height","item_index","s3_key","url","expires_at"],"title":"ArtifactOut","type":"object"},"Chain":{"properties":{"upscale":{"anyOf":[{"$ref":"#/components/schemas/ChainUpscale"},{"type":"null"}]}},"title":"Chain","type":"object"},"ChainUpscale":{"properties":{"scale":{"default":2,"description":"Upscale factor (2 or 4)","maximum":4.0,"minimum":2.0,"title":"Scale","type":"integer"}},"title":"ChainUpscale","type":"object"},"ErrorResponse":{"properties":{"code":{"title":"Code","type":"string"},"correlation_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Correlation Id"},"details":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Details"},"message":{"title":"Message","type":"string"}},"required":["code","message"],"title":"ErrorResponse","type":"object"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"title":"Detail","type":"array"}},"title":"HTTPValidationError","type":"object"},"JobCreateRequest":{"properties":{"chain":{"anyOf":[{"$ref":"#/components/schemas/Chain"},{"type":"null"}]},"count":{"default":1,"maximum":100.0,"minimum":1.0,"title":"Count","type":"integer"},"embed_metadata":{"default":true,"title":"Embed Metadata","type":"boolean"},"format":{"default":"png","title":"Format","type":"string"},"guidance":{"default":7.0,"title":"Guidance","type":"number"},"height":{"default":1024,"title":"Height","type":"integer"},"model_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model Id"},"negative_prompt":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Negative Prompt"},"prompt":{"title":"Prompt","type":"string"},"scheduler":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Scheduler"},"seed":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Seed"},"steps":{"default":30,"title":"Steps","type":"integer"},"type":{"pattern":"^generate$","title":"Type","type":"string"},"width":{"default":1024,"title":"Width","type":"integer"}},"required":["type","prompt"],"title":"JobCreateRequest","type":"object"},"JobCreated":{"properties":{"created_at":{"title":"Created At","type":"string"},"id":{"title":"Id","type":"string"},"status":{"title":"Status","type":"string"},"type":{"title":"Type","type":"string"}},"required":["id","status","type","created_at"],"title":"JobCreated","type":"object"},"JobCreatedResponse":{"properties":{"job":{"$ref":"#/components/schemas/JobCreated"}},"required":["job"],"title":"JobCreatedResponse","type":"object"},"JobStatusResponse":{"properties":{"created_at":{"title":"Created At","type":"string"},"error_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Code"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"},"id":{"title":"Id","type":"string"},"status":{"title":"Status","type":"string"},"steps":{"default":[],"items":{"$ref":"#/components/schemas/StepSummary"},"title":"Steps","type":"array"},"summary":{"additionalProperties":true,"default":{},"title":"Summary","type":"object"},"type":{"title":"Type","type":"string"},"updated_at":{"title":"Updated At","type":"string"}},"required":["id","type","status","created_at","updated_at"],"title":"JobStatusResponse","type":"object"},"ModelDescriptor":{"properties":{"capabilities":{"items":{"type":"string"},"title":"Capabilities","type":"array"},"enabled":{"default":true,"title":"Enabled","type":"boolean"},"files_json":{"items":{"additionalProperties":true,"type":"object"},"title":"Files Json","type":"array"},"id":{"title":"Id","type":"string"},"installed":{"default":false,"title":"Installed","type":"boolean"},"kind":{"title":"Kind","type":"string"},"local_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Local Path"},"name":{"title":"Name","type":"string"},"parameters_schema":{"additionalProperties":true,"title":"Parameters Schema","type":"object"},"source_uri":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source Uri"},"version":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Version"}},"required":["id","name","kind"],"title":"ModelDescriptor","type":"object"},"ModelListResponse":{"properties":{"models":{"items":{"$ref":"#/components/schemas/ModelSummary"},"title":"Models","type":"array"}},"title":"ModelListResponse","type":"object"},"ModelSummary":{"properties":{"enabled":{"default":true,"title":"Enabled","type":"boolean"},"id":{"title":"Id","type":"string"},"installed":{"default":false,"title":"Installed","type":"boolean"},"kind":{"title":"Kind","type":"string"},"name":{"title":"Name","type":"string"},"parameters_schema":{"additionalProperties":true,"title":"Parameters Schema","type":"object"},"version":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Version"}},"required":["id","name","kind"],"title":"ModelSummary","type":"object"},"ProgressItem":{"properties":{"item_index":{"title":"Item Index","type":"integer"},"progress":{"title":"Progress","type":"number"}},"required":["item_index","progress"],"title":"ProgressItem","type":"object"},"ProgressResponse":{"properties":{"items":{"default":[],"items":{"$ref":"#/components/schemas/ProgressItem"},"title":"Items","type":"array"},"progress":{"title":"Progress","type":"number"},"stages":{"default":[],"items":{"additionalProperties":true,"type":"object"},"title":"Stages","type":"array"}},"required":["progress"],"title":"ProgressResponse","type":"object"},"StepSummary":{"properties":{"name":{"title":"Name","type":"string"},"status":{"title":"Status","type":"string"}},"required":["name","status"],"title":"StepSummary","type":"object"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"title":"Location","type":"array"},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}},"required":["loc","msg","type"],"title":"ValidationError","type":"object"}}},"info":{"title":"Dream Forge API","version":"0.4.0-mvp"},"openapi":"3.1.0","paths":{"/healthz":{"get":{"operationId":"healthz_healthz_get","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Response Healthz Healthz Get","type":"object"}}},"description":"Successful Response"}},"summary":"Healthz"}},"/metrics":{"get":{"operationId":"metrics_metrics_get","responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"}},"summary":"Metrics"}},"/readyz":{"get":{"operationId":"readyz_readyz_get","responses":{"200":{"content":{"application/json":{"schema":{"title":"Response Readyz Readyz Get"}}},"description":"Successful Response"}},"summary":"Readyz"}},"/v1/":{"get":{"operationId":"root_v1__get","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"title":"Response Root V1  Get","type":"object"}}},"description":"Successful Response"}},"summary":"Root","tags":["meta"]}},"/v1/jobs":{"post":{"operationId":"create_job_v1_jobs_post","parameters":[{"in":"header","name":"Idempotency-Key","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Idempotency-Key"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobCreateRequest","examples":{"batch":{"summary":"Batch of 5 with per-item seeds","value":{"count":5,"height":64,"prompt":"m4 demo","steps":2,"type":"generate","width":64}},"single":{"summary":"Single image (default count=1)","value":{"format":"png","height":1024,"prompt":"a tranquil lake at sunrise","steps":30,"type":"generate","width":1024}}}}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobCreatedResponse"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Unprocessable Content"},"503":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Service Unavailable"}},"summary":"Create Job","tags":["jobs"]}},"/v1/jobs/{job_id}":{"get":{"operationId":"get_job_v1_jobs__job_id__get","parameters":[{"in":"path","name":"job_id","required":true,"schema":{"title":"Job Id","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobStatusResponse"}}},"description":"Successful Response"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Not Found"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Get Job","tags":["jobs"]}},"/v1/jobs/{job_id}/artifacts":{"get":{"operationId":"list_artifacts_v1_jobs__job_id__artifacts_get","parameters":[{"in":"path","name":"job_id","required":true,"schema":{"title":"Job Id","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArtifactListResponse"}}},"description":"Successful Response"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Not Found"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"List Artifacts","tags":["artifacts"]}},"/v1/jobs/{job_id}/logs":{"get":{"operationId":"get_logs_v1_jobs__job_id__logs_get","parameters":[{"in":"path","name":"job_id","required":true,"schema":{"title":"Job Id","type":"string"}},{"in":"query","name":"tail","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Tail"}},{"in":"query","name":"since_ts","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Since Ts"}}],"responses":{"200":{"content":{"application/json":{"schema":{}},"application/x-ndjson":{"examples":{"ndjson":{"summary":"Two log lines (step + artifact)","value":"{\"ts\":\"2025-09-12T21:20:00Z\",\"level\":\"info\",\"code\":\"step.start\",\"message\":\"step.start\",\"job_id\":\"<uuid>\",\"step_id\":\"<uuid>\"}\n{\"ts\":\"2025-09-12T21:20:01Z\",\"level\":\"info\",\"code\":\"artifact.written\",\"message\":\"artifact.written\",\"job_id\":\"<uuid>\",\"step_id\":\"<uuid>\",\"item_index\":0}\n"}}}},"description":"Successful Response"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Not Found"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Unprocessable Content"}},"summary":"Get Logs","tags":["logs"]}},"/v1/jobs/{job_id}/progress":{"get":{"operationId":"get_progress_v1_jobs__job_id__progress_get","parameters":[{"in":"path","name":"job_id","required":true,"schema":{"title":"Job Id","type":"string"}}],"responses":{"200":{"content":{"application/json":{"examples":{"batchProgress":{"summary":"Aggregate + per-item snapshot","value":{"items":[{"item_index":0,"progress":1.0},{"item_index":1,"progress":1.0},{"item_index":2,"progress":0.0}],"progress":0.6,"stages":[{"name":"queued_to_start","weight":0.1},{"name":"sampling","weight":0.8},{"name":"finalize","weight":0.1}]}}},"schema":{"$ref":"#/components/schemas/ProgressResponse"}}},"description":"Successful Response"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Not Found"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Get Progress","tags":["progress"]}},"/v1/jobs/{job_id}/progress/stream":{"get":{"operationId":"stream_progress_v1_jobs__job_id__progress_stream_get","parameters":[{"in":"path","name":"job_id","required":true,"schema":{"title":"Job Id","type":"string"}},{"in":"query","name":"since_ts","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Since Ts"}}],"responses":{"200":{"content":{"application/json":{"schema":{}},"text/event-stream":{"examples":{"sseExample":{"summary":"SSE progress and artifact events","value":"event: progress\ndata: {\"progress\":0.4,\"items\":[{\"item_index\":0,\"progress\":1.0},{\"item_index\":1,\"progress\":0.0}]}\n\nevent: artifact\ndata: {\"item_index\":0,\"s3_key\":\"dreamforge/..._0_64x64_123456.png\",\"format\":\"png\",\"width\":64,\"height\":64,\"seed\":123456}\n\n"}}}},"description":"Successful Response"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Not Found"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Stream Progress","tags":["progress"]}},"/v1/models":{"get":{"operationId":"list_models_v1_models_get","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelListResponse"}}},"description":"Successful Response"}},"summary":"List Models","tags":["models"]}},"/v1/models/{model_id}":{"get":{"operationId":"get_model_v1_models__model_id__get","parameters":[{"in":"path","name":"model_id","required":true,"schema":{"title":"Model Id","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelDescriptor"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Get Model","tags":["models"]}}}}